Re-work windows svc functionality
Move windows service functionality out of main, and into an svc subcommand.

Remove the service management stuff. That’s better handled by the packaging.

Replace the auto detection stuff with an explicit `svc` or `svc-debug`. The auto-detect is flakey and we should be explicit.
